Only change card order if it has changed
Fixes an issue when then a card is dragged and dropped back in its original location.

Previously this would cause the card to swap positions with the card below it.

Now the card stays in its current position and does not dispatch an update event to the phoenix backend.

The behavior when dragging to a different location within the same list, or dragging to another list remains the same.