Fix index position when dragging cards down
While working on the other PRs, I kept noticing something was slightly off when I would drag cards around in the same list.

When I looked closer, I noticed moving a card towards the top of the list always worked as expected. But when I moved a card towards the bottom of the list it would end up one position lower than the intended target.

The fix was to calculate the insert position based on its current position in the list. If the card is currently below the target, the target index will work. If the card is currently above the target, we need to subtract one from the target index.

This is because when the card moves from above to its new location, it leaves a hole in the list where it used to be. If we don't account for the "missing" card, then the card will end up one position lower than its supposed to be.

I had fun trying to reason this out. Let me know if this is still unclear, it's easier to see in code than to explain in words!

This fixes the last issue I could find with the drag behavior. Thanks for being receptive to fixes, it's nice to be able to give back to great projects :smiley: